feat(support/helper/arr): added Filter helper
Refs #6.0.x
